requirements.txt
loguru
networkx
pydantic
pyyaml
scipy
sqlglot
tree-sitter
tree-sitter-python
